refactor: move types to primary
