+++ __tests__/integration/weather.controller.spec.ts
@@ -1,5 +1,5 @@
 import { ctx } from '__tests__/setup-integration-context.js';
-import { HttpStatus } from 'src/secondary/types/http.types.js';
+import { HttpStatus } from 'src/primary/types/http.types.js';
 import { MockInstance } from 'vitest';
 import { Exception } from 'src/domain/entities/exception.entity.js';
 import { Server } from 'src/primary/adapters/server.js';
